Try moving global-styles-engine into the styles-engine package
