+++ packages/edit-site/tsconfig.json
@@ -21,7 +21,6 @@
 		{ "path": "../element" },
 		{ "path": "../escape-html" },
 		{ "path": "../fields" },
-		{ "path": "../global-styles-engine" },
 		{ "path": "../hooks" },
 		{ "path": "../html-entities" },
 		{ "path": "../i18n" },
