+++ packages/style-engine/package.json
@@ -58,7 +58,13 @@
 	},
 	"sideEffects": false,
 	"dependencies": {
-		"change-case": "^4.1.2"
+		"@wordpress/blocks": "file:../blocks",
+		"@wordpress/data": "file:../data",
+		"@wordpress/i18n": "file:../i18n",
+		"change-case": "^4.1.2",
+		"colord": "^2.9.2",
+		"deepmerge": "^4.3.0",
+		"is-plain-object": "^5.0.0"
 	},
 	"publishConfig": {
 		"access": "public"
