+++ packages/style-engine/src/utils/background.ts
