+++ packages/style-engine/src/utils/string.ts
