+++ packages/style-engine/tsconfig.json
@@ -1,4 +1,8 @@
 {
 	"$schema": "https://json.schemastore.org/tsconfig.json",
-	"extends": "../../tsconfig.base.json"
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"checkJs": false
+	},
+	"references": [ { "path": "../data" }, { "path": "../i18n" } ]
 }
