+++ tsconfig.json
@@ -28,7 +28,6 @@
 		{ "path": "packages/escape-html" },
 		{ "path": "packages/eslint-plugin" },
 		{ "path": "packages/fields" },
-		{ "path": "packages/global-styles-engine" },
 		{ "path": "packages/hooks" },
 		{ "path": "packages/html-entities" },
 		{ "path": "packages/i18n" },
